move from `_base.py` to `_post_coinit/misc.py`
